Starting to put data into postgres
This is a huge victory!!!

+++ schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
-    id int PRIMARY KEY UNIQUE NOT NULL,
+    id bigint PRIMARY KEY UNIQUE NOT NULL,
     name text NOT NULL
 );
